@@ -8,7 +8,7 @@ publish_to: "none" # Remove this line if you wish to publish to pub.dev
 version: 1.1.2+6
 
 environment:
-  sdk: ">=3.0.0 <4.0.0"
+  sdk: ">=3.4.0 <4.0.0"
 
 dependencies:
   flutter:
@@ -17,7 +17,7 @@ dependencies:
   # Util
   flutter_riverpod: ^2.1.1
   auto_route: ^5.0.4
-  intl: ^0.18.0
+  intl: ^0.19.0
   hive: ^2.2.3
   # flutter_cache_manager_hive:
   #   git:
@@ -28,14 +28,14 @@ dependencies:
   dio_cache_interceptor_hive_store: ^3.2.2
   url_launcher: ^6.1.4
   another_flushbar: ^1.12.30
-  share_plus: ^4.0.8
+  share_plus: ^7.2.2
   email_validator: ^2.0.1
   logging: ^1.0.0
-  receive_sharing_intent: ^1.4.5
+  receive_sharing_intent: ^1.5.4
   image_picker: ^0.7.3
   path_provider: ^2.0.1
-  sign_in_with_apple: ^3.2.0
-  crypto: ^3.0.1
+  sign_in_with_apple: ^4.3.0
+  crypto: ^3.0.5
   easy_localization: ^3.0.0
   flutter_localized_locales: ^2.0.3
   # metadata_fetch: ^0.4.1
@@ -45,29 +45,29 @@ dependencies:
       ref: v0.4.3
   flutter_file_dialog: ^2.3.0
   keyboard_service: ^2.0.1
-  flutter_native_image: ^0.0.6+1
+  flutter_image_compress: ^2.3.0
   version: ^2.0.0
   package_info_plus: ^4.1.0
   in_app_update: ^4.0.1 # Android Updates
   update_available: ^3.0.0 # iOS Updates
   envied: ^0.3.0+3
-  purchases_flutter: ^5.8.0
+  purchases_flutter: ^8.1.1
   permission_handler: ^10.2.0
   flutter_phoenix: ^1.1.0
   clipboard: ^0.1.3
-  connectivity_plus: ^3.0.3
-  internet_connection_checker: ^1.0.0+1
+  connectivity_plus: ^6.0.5
+  internet_connection_checker: ^2.0.0
   uni_links: ^0.5.1
   keep_screen_on: ^2.1.0
 
   # Firebase
-  cloud_firestore: ^4.3.1
-  firebase_analytics: ^10.1.0
-  firebase_app_check: ^0.2.0+1
-  firebase_auth: ^4.2.10
-  firebase_core: ^2.4.1
-  firebase_crashlytics: ^3.0.11
-  firebase_storage: ^11.0.10
+  cloud_firestore: ^5.2.1
+  firebase_analytics: ^11.2.1
+  firebase_app_check: ^0.3.0+4
+  firebase_auth: ^5.1.4
+  firebase_core: ^3.3.0
+  firebase_crashlytics: ^4.0.4
+  firebase_storage: ^12.1.3
 
   # Widgets
   flutter_snake_navigationbar: ^0.6.1
@@ -77,9 +77,8 @@ dependencies:
   adaptive_dialog: ^1.9.0
   flutter_markdown:
   auto_size_text:
-  flutter_slidable: ^0.6.0-nullsafety.0
   # flutter_toggle_tab: ^0.0.5+4
-  badges: ^2.0.0-nullsafety.1
+  badges: ^3.1.2
   # TODO: change when merged https://github.com/jamesblasco/modal_bottom_sheet/issues/325
   modal_bottom_sheet:
     git:
@@ -92,7 +91,7 @@ dependencies:
   dotted_border: ^2.0.0+2
   in_app_review: ^2.0.4
   like_button: ^2.0.5
-  blur: ^3.1.0
+  blur: ^4.0.0
   sticky_headers: ^0.3.0+2
   # fast_cached_network_image: ^1.2.0
   fast_cached_network_image:
